pyproject.toml
[tool.poetry]
name = "classification"
version = "0.1.0"
description = "Machine Learning Model Deployment for Iris Classification"
authors = ["Soyoung Yoon <soyyoon16@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
scikit-learn = "^1.5.2"
umap-learn = "^0.5.6"
plotly = "^5.24.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
sphinx = "<5.0"
sphinx-autoapi = "<2.0"
sphinx-rtd-theme = "<1.0.0"
pre-commit = "^3.8.0"
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
